+++ examples/animates/circle.sine.wave.py
@@ -5,21 +5,13 @@ From "Math Adventures with Python" Part II, Chapter 6, "Making Sine Waves"
 """
 from easygraphics import *
 import math
-import gc
-import sys
 
-
-def except_hook(cls, exception, traceback):
-    print(cls, exception, traceback)
-    sys.__excepthook__(cls, exception, traceback)
-
-
-sys.excepthook = except_hook
-
-width = 600
+width = 800
 height = 400
 
 init_graph(width, height)
+
+set_render_mode(RenderMode.RENDER_MANUAL)
 
 translate(150, height // 2)
 
@@ -29,8 +21,7 @@ r2 = 5  # radius of the point on the circle outline
 t = 0
 while is_run():
     t += 0.05
-    if delay_jfps(120):
-        gc.collect(0)
+    if delay_jfps(60):
         clear_device()
         set_color("black")
         circle(0, 0, r1)  # draw circle
@@ -47,7 +38,7 @@ while is_run():
         line(x, y, 200, y)
         fill_circle(200, y, r2)
 
-        circle_list = [y] + circle_list[:250]
+        circle_list = [y] + circle_list[:500]
 
         for i in range(1, len(circle_list)):
             line(200 + i - 1, circle_list[i - 1], 200 + i, circle_list[i])