+++ HISTORY.rst
@@ -5,6 +5,7 @@ History
 0.9.21
 -------------
 * add: show_lists_table() to display data lists in table
+* add: get_transform()/set_transform()/push_transform()/pop_transform()
 
 0.9.20
 -------------